# frontend/lyric-sync-app/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ChordKTV API",
    "version": "v1"
  },
  "paths": {
...
    "/api/database/users/{email}": {
      "get": {
        "tags": [
          "Song"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google": {
      "post": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Album": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isSingle": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "artist": {
            "type": "string",
            "nullable": true
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GeniusMetaData": {
        "type": "object",
        "properties": {
          "geniusId": {
            "type": "integer",
            "format": "int32"
          },
          "headerImageUrl": {
            "type": "string",
            "nullable": true
          },
          "headerImageThumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "songImageUrl": {
            "type": "string",
            "nullable": true
          },
          "songImageThumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          }
        },
        "additionalProperties": false
      },
...
      "LanguageCode": {
        "enum": [
          "UNK",
          "AF",
          "AR",
          "BG",
          "BN",
          "CA",
          "CS",
          "DA",
          "DE",
          "EL",
          "EN",
          "ES",
          "ET",
          "FA",
          "FI",
          "FR",
          "GU",
          "HE",
          "HI",
          "HR",
          "HU",
          "ID",
          "IT",
          "JA",
          "KO",
          "LT",
          "LV",
          "MS",
          "NL",
          "NO",
          "PL",
          "PT",
          "RO",
          "RU",
          "SK",
          "SL",
          "SR",
          "SV",
          "TA",
          "TE",
          "TH",
          "TR",
          "UK",
          "VI",
          "ZH"
        ],
        "type": "string"
      },
      "LearnedWord": {
        "required": [
          "word"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "word": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          },
          "dateLearned": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Quiz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "songId": {
            "type": "string",
            "format": "uuid"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "difficulty": {
            "type": "integer",
            "format": "int32"
          },
          "numQuestions": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "QuizOption": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "question": {
            "$ref": "#/components/schemas/QuizQuestion"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "isCorrect": {
            "type": "boolean"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "QuizQuestion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quizId": {
            "type": "string",
            "format": "uuid"
          },
          "quiz": {
            "$ref": "#/components/schemas/Quiz"
          },
          "questionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "lyricPhrase": {
            "type": "string",
            "nullable": true
          },
          "startTimestamp": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "endTimestamp": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Song": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "alternateTitles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "artist": {
            "type": "string",
            "nullable": true
          },
          "featuredArtists": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "albums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Album"
            },
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "plainLyrics": {
            "type": "string",
            "nullable": true
          },
          "lrcLyrics": {
            "type": "string",
            "nullable": true
          },
          "lrcRomanizedLyrics": {
            "type": "string",
            "nullable": true
          },
          "lrcTranslatedLyrics": {
            "type": "string",
            "nullable": true
          },
          "lrcId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "romLrcId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "youtubeId": {
            "type": "string",
            "nullable": true
          },
          "youtubeInstrumentalId": {
            "type": "string",
            "nullable": true
          },
          "alternateYoutubeIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "geniusMetaData": {
            "$ref": "#/components/schemas/GeniusMetaData"
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "required": [
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "playlistActivities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPlaylistActivity"
            },
            "nullable": true
          },
          "quizResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserQuizResult"
            },
            "nullable": true
          },
          "songActivities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSongActivity"
            },
            "nullable": true
          },
          "handwritingResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserHandwritingResult"
            },
            "nullable": true
          },
          "learnedWords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LearnedWord"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserHandwritingResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "wordTested": {
            "type": "string",
            "nullable": true
          },
          "dateCompleted": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "UserPlaylistActivity": {
        "required": [
          "playlistId",
          "playlistThumbnailUrl",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "playlistId": {
            "type": "string",
            "nullable": true
          },
          "playlistThumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "datesPlayed": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "nullable": true
          },
          "lastPlayed": {
            "type": "string",
            "format": "date-time"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "dateFavorited": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserQuizResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quizId": {
            "type": "string",
            "format": "uuid"
          },
          "quiz": {
            "$ref": "#/components/schemas/Quiz"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "language": {
            "$ref": "#/components/schemas/LanguageCode"
          },
          "dateCompleted": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "UserSongActivity": {
        "required": [
          "song"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "datesPlayed": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "nullable": true
          },
          "lastPlayed": {
            "type": "string",
            "format": "date-time"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "dateFavorited": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}